feat(ipc): wire merge:progress events from worker to renderer
