+++ src/preload/api-types.ts
@@ -318,6 +318,14 @@ export interface MergeResult {
   targetOnly: number
 }
 
+export type MergeProgress =
+  | { phase: 'parsing' }
+  | { phase: 'loading-map' }
+  | { phase: 'classifying' }
+  | { phase: 'writing'; processed: number; total: number }
+  | { phase: 'done'; result: MergeResult }
+  | { phase: 'error'; message: string }
+
 export interface MergeApi {
   prepareInput(params: { inputPath: string }): Promise<PreparedTranslationInput>
   discardInput(params: { importId: string }): Promise<{ success: boolean }>
@@ -330,6 +338,7 @@ export interface MergeApi {
     targetLang: string
     modName: string
   }): Promise<MergeResult>
+  onProgress(cb: (data: MergeProgress) => void): UnsubscribeFn
 }
 
 export interface XmlApi {
